ai-service/requirements.txt
fastapi
httpx
openai
protobuf
pydantic
pydantic_settings
python-dotenv
SQLAlchemy
tenacity
uvicorn
